backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic[email]==2.7.1
google-auth==2.27.0
tensorflow==2.16.1
librosa==0.10.2
numpy==1.26.4
scikit-learn==1.4.2
soundfile==0.12.1
